Developer

+++ app/Cache.php
@@ -4,8 +4,8 @@
  *
  * @package App
  *
- * @copyright YetiForce Sp. z o.o
- * @license   YetiForce Public License 3.0 (licenses/LicenseEN.txt or yetiforce.com)
+ * @copyright YetiForce S.A.
+ * @license   YetiForce Public License 5.0 (licenses/LicenseEN.txt or yetiforce.com)
  * @author    Mariusz Krzaczkowski <m.krzaczkowski@yetiforce.com>
  * @author    Radosław Skrzypczak <r.skrzypczak@yetiforce.com>
  */
@@ -17,8 +17,13 @@ namespace App;
  */
 class Cache
 {
+	/** @var int Long time data storage */
 	const LONG = 3600;
+
+	/** @var int Medium time data storage */
 	const MEDIUM = 300;
+
+	/** @var int Short time data storage */
 	const SHORT = 60;
 	public static $pool;
 	public static $staticPool;
@@ -53,7 +58,7 @@ class Cache
 	 *
 	 * @return mixed
 	 */
-	public static function get($nameSpace, $key)
+	public static function get(string $nameSpace, string $key)
 	{
 		return static::$pool->get("$nameSpace-$key");
 	}
@@ -66,7 +71,7 @@ class Cache
 	 *
 	 * @return bool
 	 */
-	public static function has($nameSpace, $key): bool
+	public static function has(string $nameSpace, string $key): bool
 	{
 		return static::$pool->has("$nameSpace-$key");
 	}
@@ -81,7 +86,7 @@ class Cache
 	 *
 	 * @return bool
 	 */
-	public static function save($nameSpace, $key, $value = null, $duration = self::MEDIUM)
+	public static function save(string $nameSpace, string $key, $value = null, $duration = self::MEDIUM)
 	{
 		if (!static::$pool->save("$nameSpace-$key", $value, $duration)) {
 			Log::warning("Error writing to cache. Key: $nameSpace-$key | Value: " . var_export($value, true));
@@ -97,7 +102,7 @@ class Cache
 	 *
 	 * @return bool
 	 */
-	public static function delete($nameSpace, $key)
+	public static function delete(string $nameSpace, string $key)
 	{
 		static::$pool->delete("$nameSpace-$key");
 	}
@@ -107,9 +112,9 @@ class Cache
 	 *
 	 * @return bool
 	 */
-	public static function clear()
+	public static function clear(): bool
 	{
-		static::$pool->clear();
+		return static::$pool->clear();
 	}
 
 	/**
@@ -120,7 +125,7 @@ class Cache
 	 *
 	 * @return mixed
 	 */
-	public static function staticGet($nameSpace, $key)
+	public static function staticGet(string $nameSpace, string $key = '')
 	{
 		return static::$staticPool->get("$nameSpace-$key");
 	}
@@ -133,7 +138,7 @@ class Cache
 	 *
 	 * @return bool
 	 */
-	public static function staticHas($nameSpace, $key = '')
+	public static function staticHas(string $nameSpace, string $key = '')
 	{
 		return static::$staticPool->has("$nameSpace-$key");
 	}
@@ -148,7 +153,7 @@ class Cache
 	 *
 	 * @return bool
 	 */
-	public static function staticSave($nameSpace, $key, $value = null)
+	public static function staticSave(string $nameSpace, string $key, $value = null)
 	{
 		return static::$staticPool->save("$nameSpace-$key", $value);
 	}
@@ -161,7 +166,7 @@ class Cache
 	 *
 	 * @return bool
 	 */
-	public static function staticDelete($nameSpace, $key)
+	public static function staticDelete(string $nameSpace, string $key)
 	{
 		static::$staticPool->delete("$nameSpace-$key");
 	}
@@ -187,7 +192,11 @@ class Cache
 			return false;
 		}
 		register_shutdown_function(function () {
-			static::resetOpcache();
+			try {
+				static::resetOpcache();
+			} catch (\Throwable $e) {
+				\App\Log::error($e->getMessage() . PHP_EOL . $e->__toString());
+			}
 		});
 		return static::$clearOpcache = true;
 	}
@@ -243,6 +252,9 @@ class Cache
 		$exclusion = ['.htaccess', 'index.html'];
 		$s = $i = 0;
 		foreach (['pdf', 'import', 'mail', 'vtlib', 'rss_cache', 'upload', 'templates_c'] as $dir) {
+			if (!file_exists(ROOT_DIRECTORY . "/cache/{$dir}")) {
+				continue;
+			}
 			foreach (new \RecursiveIteratorIterator(new \RecursiveDirectoryIterator(ROOT_DIRECTORY . "/cache/{$dir}", \RecursiveDirectoryIterator::SKIP_DOTS), \RecursiveIteratorIterator::SELF_FIRST) as $item) {
 				if ($item->isFile() && !\in_array($item->getBasename(), $exclusion) && $item->getMTime() < $time && $item->getATime() < $time) {
 					$s += $item->getSize();
@@ -253,7 +265,7 @@ class Cache
 		}
 		foreach ([ROOT_DIRECTORY . '/cache', \App\Fields\File::getTmpPath()] as $dir) {
 			foreach ((new \DirectoryIterator($dir)) as $item) {
-				if ($item->isFile() && 'index.html' !== $item->getBasename()) {
+				if ($item->isFile() && 'index.html' !== $item->getBasename() && $item->getMTime() < $time && $item->getATime() < $time) {
 					$s += $item->getSize();
 					unlink($item->getPathname());
 					++$i;
