+++ modules/Settings/Roles/models/Record.php
@@ -6,7 +6,7 @@
  * The Initial Developer of the Original Code is vtiger.
  * Portions created by vtiger are Copyright (C) vtiger.
  * All Rights Reserved.
- * Contributor(s): YetiForce Sp. z o.o
+ * Contributor(s): YetiForce S.A.
  * *********************************************************************************** */
 
 /**
@@ -14,6 +14,12 @@
  */
 class Settings_Roles_Record_Model extends Settings_Vtiger_Record_Model
 {
+
+	/** @var array Stores child roles. */
+	private $children;
+	/** @var array All the profiles associated with the current role. */
+	private $profiles;
+
 	/**
 	 * Function to get the Id.
 	 *
@@ -355,12 +361,12 @@ class Settings_Roles_Record_Model extends Settings_Vtiger_Record_Model
 		$profileIds = $this->get('profileIds');
 		if ($rolePreviousData) {
 			$oldProfileIds = $this->getProfileIdList();
-			if ($rolePreviousData['listrelatedrecord'] != $this->get('listrelatedrecord') ||
-					$rolePreviousData['previewrelatedrecord'] != $this->get('previewrelatedrecord') ||
-					$rolePreviousData['editrelatedrecord'] != $this->get('editrelatedrecord') ||
-					$rolePreviousData['permissionsrelatedfield'] != $permissionsRelatedField ||
-					$rolePreviousData['searchunpriv'] != $searchunpriv ||
-				($profileIds && !empty(array_merge(array_diff($profileIds, $oldProfileIds), array_diff($oldProfileIds, $profileIds))))) {
+			if ($rolePreviousData['listrelatedrecord'] != $this->get('listrelatedrecord')
+					|| $rolePreviousData['previewrelatedrecord'] != $this->get('previewrelatedrecord')
+					|| $rolePreviousData['editrelatedrecord'] != $this->get('editrelatedrecord')
+					|| $rolePreviousData['permissionsrelatedfield'] != $permissionsRelatedField
+					|| $rolePreviousData['searchunpriv'] != $searchunpriv
+				|| ($profileIds && !empty(array_merge(array_diff($profileIds, $oldProfileIds), array_diff($oldProfileIds, $profileIds))))) {
 				\App\Privilege::setAllUpdater();
 			}
 		}
@@ -384,7 +390,10 @@ class Settings_Roles_Record_Model extends Settings_Vtiger_Record_Model
 		\App\Cache::delete(__CLASS__, $roleId);
 		\App\Cache::delete('RoleDetail', $roleId);
 		\App\Cache::delete('getUsersByCompany', '');
-		\App\Cache::delete('getUsersByCompany', $this->get('company'));
+		if ($this->get('company')) {
+			\App\Cache::delete('getUsersByCompany', $this->get('company'));
+		}
+		\App\Cache::delete('getCompanyRoles', '');
 		if (isset($rolePreviousData['company'])) {
 			\App\Cache::delete('getUsersByCompany', $rolePreviousData['company']);
 		}
@@ -428,12 +437,8 @@ class Settings_Roles_Record_Model extends Settings_Vtiger_Record_Model
 		}
 	}
 
-	/**
-	 * Function to get the list view actions for the record.
-	 *
-	 * @return <Array> - Associate array of Vtiger_Link_Model instances
-	 */
-	public function getRecordLinks()
+	/** {@inheritdoc} */
+	public function getRecordLinks(): array
 	{
 		$links = [];
 		if ($this->getParent()) {
@@ -441,7 +446,7 @@ class Settings_Roles_Record_Model extends Settings_Vtiger_Record_Model
 				[
 					'linktype' => 'LISTVIEWRECORD',
 					'linklabel' => 'LBL_EDIT_RECORD',
-					'linkurl' => $this->getListViewEditUrl(),
+					'linkurl' => $this->getEditViewUrl(),
 					'linkicon' => 'yfi yfi-full-editing-view',
 				],
 				[
@@ -480,7 +485,6 @@ class Settings_Roles_Record_Model extends Settings_Vtiger_Record_Model
 			$roles[$role->getId()] = $role;
 		}
 		$dataReader->close();
-
 		return $roles;
 	}
